# apps/private-paid-order-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Private Paid-Order Backend
  version: 2.6.0
  description: >-
    Phase 1 private backend. WordPress may call only the signed paid-order intake endpoint.
    Operator routes require an authenticated server-side session. Proprietary internal content is never exposed.
servers:
  - url: https://PLACEHOLDER_STAGING_BACKEND_HOST
    description: Staging; replace before handoff
paths:
...
  /api/v1/webhooks/wordpress/paid-orders:
    post:
      summary: Accept one successfully paid WordPress/WooCommerce order
      description: >-
        HMAC signs the exact raw JSON bytes using timestamp + '.' + nonce + '.' + rawBody.
        Requires fresh timestamp/nonce and a stable logical idempotency key.
      parameters:
        - $ref: '#/components/parameters/WebhookKeyId'
        - $ref: '#/components/parameters/WebhookTimestamp'
        - $ref: '#/components/parameters/WebhookNonce'
        - $ref: '#/components/parameters/WebhookSignature'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./paid-order.schema.json
      responses:
        '201':
          description: New order/job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOrder'
        '200':
          description: Safe duplicate; existing order/job returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOrder'
        '400':
          $ref: '#/components/responses/Refused'
        '401':
          $ref: '#/components/responses/Refused'
        '409':
          $ref: '#/components/responses/Refused'
        '422':
          description: Field/catalogue/pricing parity validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /api/v1/jobs/{jobId}/history:
    get:
      summary: Operator-safe job status history
      security: [{ operatorSession: [] }]
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200': { description: Status history }
        '401': { description: Authentication required }
        '404': { description: Job not found }
...
components:
  securitySchemes:
    operatorSession:
      type: apiKey
      in: cookie
      name: __Host-mayne_session
      description: Opaque production session cookie; database stores only its hash.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Must also match the CSRF cookie and server-side hash.
  parameters:
    WebhookKeyId:
      in: header
      name: X-Webhook-Key-Id
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
    WebhookTimestamp:
      in: header
      name: X-Webhook-Timestamp
      required: true
      schema: { type: string, pattern: '^[0-9]+$' }
    WebhookNonce:
      in: header
      name: X-Webhook-Nonce
      required: true
      schema: { type: string, minLength: 16, maxLength: 200 }
    WebhookSignature:
      in: header
      name: X-Webhook-Signature
      required: true
      schema: { type: string, pattern: '^sha256=[a-f0-9]{64}$' }
    IdempotencyKey:
      in: header
      name: X-Idempotency-Key
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
    JobId:
      in: path
      name: jobId
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    AcceptedOrder:
      type: object
      required: [success, requestId, orderId, jobId, publicJobNumber, status, duplicate]
      properties:
        success: { const: true }
        requestId: { type: string }
        orderId: { type: string }
        jobId: { type: string }
        publicJobNumber: { type: string }
        status: { const: accepted }
        duplicate: { type: boolean }
    FieldError:
      type: object
      required: [path, code, message]
      properties:
        path: { type: string }
        code: { type: string }
        message: { type: string }
    ValidationFailure:
      type: object
      required: [success, requestId, error]
      properties:
        success: { const: false }
        requestId: { type: string }
        error:
          type: object
          required: [code, message, fields]
          properties:
            code: { const: PAYLOAD_VALIDATION_FAILED }
            message: { type: string }
            fields:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }
    Refusal:
      type: object
      required: [success, requestId, error]
      properties:
        success: { const: false }
        requestId: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
  responses:
    Refused:
      description: Request refused
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Refusal' }
    Unavailable:
      description: Temporary backend dependency/transaction failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Refusal' }